PR fix: move types from global.ts to separate files. Move language types to resourceadm types file
