+++ frontend/resourceadm/types/DeployError.d.ts
@@ -1,0 +1,4 @@
+export interface DeployError {
+  message: string;
+  pageWithError: 'about' | 'policy';
+}
