+++ frontend/resourceadm/types/EnvironmentType.d.ts
@@ -1,0 +1,1 @@
+export type EnvironmentType = 'AT21' | 'AT22' | 'AT23' | 'AT24' | 'TT02' | 'PROD';
