+++ frontend/resourceadm/types/NavigationBarPage.d.ts
@@ -1,0 +1,1 @@
+export type NavigationBarPage = 'about' | 'policy' | 'deploy' | 'migration';
