+++ frontend/resourceadm/types/Translation.d.ts
@@ -1,0 +1,1 @@
+export type Translation = 'none' | 'title' | 'description' | 'rightDescription';
