--- frontend/resourceadm/types/global.ts
@@ -1,14 +1,0 @@
-export type NavigationBarPage = 'about' | 'policy' | 'deploy' | 'migration';
-
-export type ValidLanguage = 'nb' | 'nn' | 'en';
-
-export type SupportedLanguage = Record<ValidLanguage, string>;
-
-export interface DeployError {
-  message: string;
-  pageWithError: 'about' | 'policy';
-}
-
-export type Translation = 'none' | 'title' | 'description' | 'rightDescription';
-
-export type EnvironmentType = 'AT21' | 'AT22' | 'AT23' | 'AT24' | 'TT02' | 'PROD';
